# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger UI",
        "description": "L5 Swagger API description",
        "contact": {
            "name": "Dinh Huan",
            "email": "dinhhuan1357@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/products/pages": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "sản phẩm quan tâm nhiều",
                "description": "Lấy sản phẩm hot nhất hiển thị trên trang chủ",
                "operationId": "2940002b2cd1dec288a466e18e8bff93",
                "responses": {
                    "200": {
                        "description": "Lấy sản phẩm thành công"
                    },
                    "400": {
                        "description": "Lỗi mạng"
                    }
                }
            }
        },
        "/api/collection/products": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "lọc sản phẩm",
                "description": "Lọc sản phẩm theo nhãn hàng (brand), giá, loại sản phẩm con",
                "operationId": "8d5266d3c56a83a952cb1ee594ffeee8",
                "parameters": [
                    {
                        "name": "params",
                        "in": "query",
                        "description": "Lọc sản phẩm với tiêu chí",
                        "required": false,
                        "schema": {
                            "type": "object",
                            "example": {
                                "sortbyprice": "asc",
                                "category": "1"
                            },
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lọc thành công"
                    },
                    "400": {
                        "description": "Lọc thất bại"
                    }
                }
            }
        },
        "/api/collection/category/{id}": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "sản phẩm của một loại (category)",
                "description": "Lấy tất cả sản phẩm theo loại sản phẩm (category)",
                "operationId": "fb68cb5415d9cd2ec126be9f20d54ce3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy sản phẩm thành công"
                    },
                    "400": {
                        "description": "Không tìm thấy sản phẩm"
                    }
                }
            }
        },
        "/api/collection/all": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "tất cả sản phẩm",
                "description": "Lấy tất cả sản phẩm",
                "operationId": "16a6b87468e0b4d2b7ddcf113d9fc061",
                "responses": {
                    "200": {
                        "description": "lấy sản phẩm thành công"
                    },
                    "400": {
                        "description": "lấy sản phẩm thất bại"
                    }
                }
            }
        },
        "/api/products/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Tạo một sản phẩm",
                "description": "Tạo sản phẩm mới gửi cùng với ảnh, lưu ý hành động này cũng sinh ra một variant mặc định tasteless",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "ProductName",
                                    "description",
                                    "Price",
                                    "Sale",
                                    "StockQuantity",
                                    "CategoryID",
                                    "BrandID",
                                    "Images[]"
                                ],
                                "properties": {
                                    "productName": {
                                        "description": "Tên sản phẩm",
                                        "format": "name",
                                        "example": "Ostrovit Micellar Casein - 700 grams"
                                    },
                                    "description": {
                                        "description": "Mô tả sản phẩm (được lưu trữ dạng html trong database)",
                                        "example": "sản phẩm số 1"
                                    },
                                    "price": {
                                        "description": "Giá gốc của sản phẩm",
                                        "format": "int32",
                                        "example": "100000"
                                    },
                                    "sale": {
                                        "description": "tỷ lệ sale sản phẩm",
                                        "format": "int32",
                                        "example": "10"
                                    },
                                    "stockQuantity": {
                                        "description": "Số lượng sản phẩm (tổng tất cả các variant)",
                                        "format": "int32",
                                        "example": "9"
                                    },
                                    "categoryId": {
                                        "description": "Mã loại sản phẩm",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "brandId": {
                                        "description": "Mã nhãn hiệu",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "image": {
                                        "description": "File ảnh của sản phẩm (4 ảnh) phải thuộc các định dạng: .jqg, .webp, .png",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo sản phẩm thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Tạo sản phẩm thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fail": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo sản phẩm thất bại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Lấy thông tin sản phẩm",
                "description": "Lấy thông tin một sản phẩm có {id}",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của sản phẩm cần tìm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Fail to get"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "xóa sản phẩm",
                "description": "xóa sản phẩm (hành động nguy hiểm)",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id sản phẩm cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "xóa sản phẩm thành công"
                    },
                    "400": {
                        "description": "xóa sản phẩm thất bại"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Product"
                ],
                "summary": "cập nhật thông tin",
                "description": "Cập nhật thông tin product cụ thể, id của product có trong path",
                "operationId": "7b6a99d3d6d5c75a1ef063b213e8d068",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của sản phẩm cần update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productName": {
                                        "type": "string",
                                        "example": "whey protein"
                                    },
                                    "shortDescription": {
                                        "type": "string",
                                        "example": "...."
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "..."
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "sale": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "stockQuantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "categoryId": {
                                        "type": "=integer",
                                        "example": 1
                                    },
                                    "brandId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật sản phẩm thành công"
                    },
                    "400": {
                        "description": "Cập nhật sản phẩm thất bại"
                    }
                }
            }
        },
        "/api/products/{id}/image": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Thêm ảnh cho sản phẩm",
                "description": "Thêm ảnh sản phẩm",
                "operationId": "4865b3c80afb0b67fd6f6d75d277f3b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "mã sản phẩm",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "Images[]": {
                                        "description": "File ảnh của sản phẩm (tùy ý)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tải ảnh lên thành công"
                    },
                    "400": {
                        "description": "Tải ảnh lên thất bại"
                    }
                }
            }
        },
...
        "/api/products/image/{image_id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Xóa ảnh sản phẩm",
                "description": "Xóa 1 ảnh sản phẩm xác định",
                "operationId": "bf039d4847ae66a2252ad1e1c243bd16",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "id của ảnh",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa ảnh thành công"
                    },
                    "400": {
                        "description": "Xóa ảnh thất bại"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}